fixed unitest
