--- cmd/main.go
@@ -1,12 +1,0 @@
-package main
-
-import "github.com/armosec/kubescape/cmd"
-
-// "github.com/armosec/kubescape/httphandler/listener"
-
-func main() {
-	cmd.Execute()
-
-	// // cmd.Execute()
-	// listener.SetupHTTPListener()
-}
